sample_app/cerebral_genai/code/rag-on-edge-cerebral/requirements.txt
Flask
Flask-Session
redis
influxdb-client
plotly
pandas
openai
python-dotenv
kaleido
